ST3 format; P8ify all colors; color pixels.
All colors specified in the Pico-8 scheme are on one of the two Pico-8 palettes. I have named all P8 colors, including the secret palette colors that are not actually used in the scheme, for the convenience of future developers who want to tinker with this.

The per-nybble colors for the sprite sheet and label are provided here, matching the highly specific scopes emitted by the syntax.

Some redundant scopes were removed. All standard scopes listed in the Sublime Text 3 docs are now provided.